Move Device Registry service classes to registry base module
The classes defining the Device Registry services have been moved to the
registry base module so that all Device Registry related classes are now
in one place.
