--- service-base/src/main/java/org/eclipse/hono/service/management/tenant/TenantBackend.java
@@ -1,23 +1,0 @@
-/*******************************************************************************
- * Copyright (c) 2019 Contributors to the Eclipse Foundation
- *
- * See the NOTICE file(s) distributed with this work for additional
- * information regarding copyright ownership.
- *
- * This program and the accompanying materials are made available under the
- * terms of the Eclipse Public License 2.0 which is available at
- * http://www.eclipse.org/legal/epl-2.0
- *
- * SPDX-License-Identifier: EPL-2.0
- *******************************************************************************/
-
-package org.eclipse.hono.service.management.tenant;
-
-import org.eclipse.hono.service.tenant.TenantService;
-
-/**
- * Unifying interface for implementing tenant backend.
- */
-public interface TenantBackend extends TenantManagementService, TenantService {
-
-}
