+++ services/device-registry-base/src/main/java/org/eclipse/hono/service/management/credentials/CommonSecret.java
