+++ services/device-registry-base/src/main/java/org/eclipse/hono/service/registration/RegistrationMessageFilter.java
