+++ services/device-registry-base/src/main/java/org/eclipse/hono/service/tenant/DelegatingTenantAmqpEndpoint.java
