+++ services/device-registry-base/src/test/java/org/eclipse/hono/service/credentials/CredentialsMessageFilterTest.java
