+++ services/device-registry-base/src/test/java/org/eclipse/hono/service/management/credentials/CredentialsTest.java
