+++ services/device-registry-base/src/test/java/org/eclipse/hono/service/management/tenant/TrustedCertificateAuthorityTest.java
