+++ services/device-registry-base/src/test/java/org/eclipse/hono/service/registration/RegistrationServiceTests.java
