+++ services/device-registry-mongodb/pom.xml
@@ -34,7 +34,7 @@
         </dependency>
         <dependency>
             <groupId>org.eclipse.hono</groupId>
-            <artifactId>hono-service-base</artifactId>
+            <artifactId>hono-service-device-registry-base</artifactId>
             <version>${project.version}</version>
             <classifier>tests</classifier>
             <type>test-jar</type>
